moved coordinate transformation to area
